move InitParserVersion
