+++ parser/pt.go
@@ -13,7 +13,6 @@ import (
 	"errors"
 	"log"
 	"net"
-	"os"
 	"path/filepath"
 	"strconv"
 	"strings"
@@ -21,28 +20,6 @@ import (
 
 	"github.com/m-lab/traceroute-caller/schema"
 )
-
-func init() {
-	InitParserVersion()
-}
-
-var gParserVersion string
-
-// InitParserVersion initializes the gParserVersion variable for use by all parsers.
-func InitParserVersion() string {
-	release, ok := os.LookupEnv("RELEASE_TAG")
-	if ok && release != "empty_tag" {
-		gParserVersion = "https://github.com/m-lab/traceroute-caller/tree/" + release
-	} else {
-		hash := os.Getenv("COMMIT_HASH")
-		if len(hash) >= 8 {
-			gParserVersion = "https://github.com/m-lab/traceroute-caller/tree/" + hash[0:8]
-		} else {
-			gParserVersion = "local development"
-		}
-	}
-	return gParserVersion
-}
 
 // PTFileName is the file name of paris-traceroute output.
 type PTFileName struct {
